+++ lib/core/pipeline.ml
@@ -43,6 +43,22 @@ let read_file_as_metadata (type a) (module P : Required.DATA_PROVIDER)
   Task.make (Deps.singleton path) (fun () ->
       Eff.read_file_as_metadata (module P) (module R) ?snapshot ~on:`Source path)
 
+let read_template (module T : Required.DATA_TEMPLATE) ?(snapshot = true)
+    ?(strict = true) template =
+  let open Task in
+  read_file ~snapshot template
+  >>> lift ~has_dynamic_dependencies:false (fun template_content ->
+          let callback (type a)
+              (module I : Required.DATA_INJECTABLE with type t = a) ~metadata
+              content =
+            let parameters =
+              ("yocaml_body", Data.string content) :: I.normalize metadata
+              |> List.map (fun (k, v) -> (k, T.from v))
+            in
+            T.render ~strict parameters template_content
+          in
+          callback)
+
 let as_template (type a) (module T : Required.DATA_TEMPLATE)
     (module I : Required.DATA_INJECTABLE with type t = a) ?(snapshot = true)
     ?(strict = true) template =