+++ CHANGES.md
@@ -1,3 +1,18 @@
+### unreleased
+
+#### Yocaml
+
+- Add `Pipeline.read_template` to have a better fit with Applicative API [gr-im](https://github.com/gr-im)
+
+#### Yocaml_jingoo
+
+- Add `read_template` to have a better fit with Applicative API [gr-im](https://github.com/gr-im)
+
+#### Yocaml_mustache
+
+- Add `read_template` to have a better fit with Applicative API [gr-im](https://github.com/gr-im)
+
+
 ### v2.4.1 2025-09-01 Nantes (France)
 
 #### Yocaml